third_party/WebKit/LayoutTests/http/tests/inspector/service-workers/user-agent-override-expected.txt

CONSOLE MESSAGE: line 78: registerServiceWorker success
CONSOLE MESSAGE: line 78: registerServiceWorker success
CONSOLE MESSAGE: line 95: unregisterServiceWorker success
Tests that User-Agent override works for requests from Service Workers.

Enable emulation and set User-Agent override
Disable emulation
Test complete
